+++ packages/frontend/core/src/mobile/components/navigation/nodes/doc/operations.tsx
@@ -80,12 +80,17 @@ export const useNavigationPanelDocNodeAddLinkedPage = (
 
 export const useNavigationPanelDocNodeOperations = (docId: string) => {
   const t = useI18n();
-  const { workbenchService, docsService, compatibleFavoriteItemsAdapter } =
-    useServices({
-      DocsService,
-      WorkbenchService,
-      CompatibleFavoriteItemsAdapter,
-    });
+  const {
+    workbenchService,
+    docsService,
+    compatibleFavoriteItemsAdapter,
+    guardService,
+  } = useServices({
+    DocsService,
+    WorkbenchService,
+    CompatibleFavoriteItemsAdapter,
+    GuardService,
+  });
 
   const { openConfirmModal } = useConfirmModal();
 
@@ -117,7 +122,11 @@ export const useNavigationPanelDocNodeOperations = (docId: string) => {
       confirmButtonOptions: {
         variant: 'error',
       },
-      onConfirm() {
+      async onConfirm() {
+        if (!(await guardService.can('Doc_Trash', docId))) {
+          toast(t['com.affine.no-permission']());
+          return;
+        }
         docRecord.moveToTrash();
         track.$.navigationPanel.docs.deleteDoc({
           control: 'button',
@@ -125,7 +134,7 @@ export const useNavigationPanelDocNodeOperations = (docId: string) => {
         toast(t['com.affine.toastMessage.movedTrash']());
       },
     });
-  }, [docRecord, openConfirmModal, t]);
+  }, [docId, docRecord, guardService, openConfirmModal, t]);
 
   const handleOpenInNewTab = useCallback(() => {
     workbenchService.workbench.openDoc(docId, {
@@ -296,43 +305,40 @@ export const useNavigationPanelDocNodeOperationsMenu = (
           </MenuItem>
         ),
       },
-      ...(!options.isInFolder
-        ? [
-            {
-              index: 9999,
-              view: <MenuSeparator key="menu-separator" />,
-            },
-            {
-              index: 10000,
-              view: options.linkedFromDocId ? (
-                <Guard docId={options.linkedFromDocId} permission="Doc_Update">
-                  {canEdit => (
-                    <MenuItem
-                      prefixIcon={<UnlinkIcon />}
-                      onClick={handleRemoveLinkedDoc}
-                      disabled={!canEdit}
-                    >
-                      {t['com.affine.rootAppSidebar.doc.remove-link']()}
-                    </MenuItem>
-                  )}
-                </Guard>
-              ) : (
-                <Guard docId={docId} permission="Doc_Trash">
-                  {canMoveToTrash => (
-                    <MenuItem
-                      type={'danger'}
-                      prefixIcon={<DeleteIcon />}
-                      onClick={handleMoveToTrash}
-                      disabled={!canMoveToTrash}
-                    >
-                      {t['com.affine.moveToTrash.title']()}
-                    </MenuItem>
-                  )}
-                </Guard>
-              ),
-            },
-          ]
-        : []),
+      {
+        index: 9999,
+        view: <MenuSeparator key="menu-separator" />,
+      },
+      {
+        index: 10000,
+        view:
+          options.linkedFromDocId && !options.isInFolder ? (
+            <Guard docId={options.linkedFromDocId} permission="Doc_Update">
+              {canEdit => (
+                <MenuItem
+                  prefixIcon={<UnlinkIcon />}
+                  onClick={handleRemoveLinkedDoc}
+                  disabled={!canEdit}
+                >
+                  {t['com.affine.rootAppSidebar.doc.remove-link']()}
+                </MenuItem>
+              )}
+            </Guard>
+          ) : (
+            <Guard docId={docId} permission="Doc_Trash">
+              {canMoveToTrash => (
+                <MenuItem
+                  type={'danger'}
+                  prefixIcon={<DeleteIcon />}
+                  onClick={handleMoveToTrash}
+                  disabled={!canMoveToTrash}
+                >
+                  {t['com.affine.moveToTrash.title']()}
+                </MenuItem>
+              )}
+            </Guard>
+          ),
+      },
     ],
     [
       docId,
